# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Authentication System API",
    "version": "1.0.0",
    "description": "API documentation for the Authentication System with Admin Roles"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/user/auth/google": {
      "post": {
        "summary": "User login with Google",
        "tags": [
          "User Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "Google ID token"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "email": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid Google token"
          }
        }
      }
    },
    "/api/user/auth/email": {
      "post": {
        "summary": "User login with email/password",
        "tags": [
          "User Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "user123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful login",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "email": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/api/user/auth/logout": {
      "post": {
        "summary": "User logout",
        "tags": [
          "User Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully logged out"
          },
          "400": {
            "description": "No token provided"
          },
          "500": {
            "description": "Logout failed"
          }
        }
      }
    },
    "/api/user/signup": {
      "post": {
        "summary": "User signup with email verification (sends OTP, stores data temporarily)",
        "tags": [
          "User Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "name"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  },
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP sent for verification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    },
                    "signupToken": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or user exists"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/user/auth/verify-otp": {
      "post": {
        "summary": "Verify OTP and create user account",
        "tags": [
          "User Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "signupToken",
                  "otp"
                ],
                "properties": {
                  "signupToken": {
                    "type": "string",
                    "description": "Signup token from signup response"
                  },
                  "otp": {
                    "type": "string",
                    "description": "OTP code"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created and verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "token": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "email": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/user/auth/resend-otp": {
      "post": {
        "summary": "Resend OTP for signup verification",
        "tags": [
          "User Authentication"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "signupToken"
                ],
                "properties": {
                  "signupToken": {
                    "type": "string",
                    "description": "Signup token from signup response"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP resent successfully"
          },
          "400": {
            "description": "Invalid or expired signup session"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}